Use signed integers for coordinates and length in drawing functions
Even if this makes introduces a little complexity in the implementation,
the use of signed integers allows user applications to write much simpler
code and just get eventually part of the pixels dropped out. For example:
by using only unsigned types, it was possible to draw a rect starting
from a valid point on-screen (or better, in the current window) and than
having part of it cut because it was off-screen BUT, in case its top-left
corner was off-screen, the whole rect was discarded, which is often an
undesired effect. Usually people want just the off-screen pixels to be
left off-screen. Now instead, a rect can be drawn at (-10, -10) for
example and still have some part of it visible on the screen (window).